fix memoize util causing cache mismatch
